Extracted `LogLine` to its own class
`Parser` and `LogLine` are now in separate files. Previously most of the code in `rp_parser.rb` belonged to `LogLine` which was rather silly.

Also fixed all references to `Parser::LogLine` as it's now a top-level class.

This commit picked up some pending refactorings to `parse_mirc.rb` that had been in-progress.