Tag implications and aliases
Added two new options to the configuration file, tag_implications and tag_aliases, each a map of tags to implied and aliased tags respectively. A valid example configuation is in the default config file. Closes #36. Closes #12.

Added tests for tag implications and aliases. Added a test RP for the same.
Added a service object to perform tag implication.

Changed Tag#to_s to return a character tag with "char:" in front.
Added Tag.[], inspired by Hash.[], to convert a list of strings into Tags.

Cleaned up the RpLogGenerator tests to take advantage of enumeration for checking if the right files were kept or skipped.