Moved RP formatting to a Liquid template
Added the rp_line.html include, which replaces most of the output
methods in LogLine. A list of all the merged lines is passed to each RP
page now, and it uses this include (a Liquid template) to generate the
actual HTML. This allows users to change the output, and ties the
structure of the output closer to HTML rather than generating it with
Ruby. Closes #57.

Part of this included removing a ton of methods from LogLine. The tests
haven't yet been updated.