Expose an executable to simplify site scaffolding
The gem now exposes an `rplogs` executable that can handle generating a
new scaffold on its own, instead of requiring users to create a Gemfile,
create a Rakefile, and run a rake task.

The new process: `rplogs --new DIR`, which creates the directory DIR,
creates a Gemfile inside it, runs bundler, and copies a new scaffold
into it. Basically combining the previous three steps into one.

Closes #63.