rplogs update command to recopy theme files
Running the `rplogs update` command from an existing site will recopy
the theme files (_sass/, _includes/, js/, css/, _layouts). This is
useful if there's an upstream update to the default theme.

This uses the same checking as the old `rake rp_logs:new`, which is to
say it asks about overwriting an existing _sass directory. This...
doesn't exactly make sense.

Progress on #69.