`rplogs update` checks for wrong directory
When a `_config.yml` isn't found, it prompts the user to continue and
informs them that this may not be a Jekyll site directory.

It still prompts for overwriting when a `_sass` directory exists.

Also reorganized the location of the gem-provided rake tasks so they
weren't nested three levels deep inside a class.

Closes #69.